// library-frontend-service/src/generated/external/api.ts
export interface BookInfo {
    /**
     * 
     * @type {BookInfoOnix}
     * @memberof BookInfo
     */
    onix?: BookInfoOnix;
    /**
     * 
     * @type {BookInfoHanmoto}
     * @memberof BookInfo
     */
    hanmoto?: BookInfoHanmoto;
    /**
     * 
     * @type {BookInfoSummary}
     * @memberof BookInfo
     */
    summary?: BookInfoSummary;
}
/**
 * 版元独自書誌
 * @export
 * @interface BookInfoHanmoto
 */
export interface BookInfoHanmoto {
    /**
     * 公開日
     * @type {Date}
     * @memberof BookInfoHanmoto
     */
    datekoukai?: Date;
    /**
     * 最終更新日
     * @type {Date}
     * @memberof BookInfoHanmoto
     */
    datemodified: Date;
    /**
     * 最終作成日
     * @type {Date}
     * @memberof BookInfoHanmoto
     */
    datecreated?: Date;
    /**
     * 出版日
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    dateshuppan?: string;
    /**
     * 読者書き込みページ数
     * @type {number}
     * @memberof BookInfoHanmoto
     */
    dokushakakikomipagesuu?: number;
    /**
     * 販売区分
     * @type {boolean}
     * @memberof BookInfoHanmoto
     */
    kubunhanbai?: boolean;
    /**
     * 本の綴じ方
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    toji?: string;
    /**
     * 在庫ステータス
     * @type {number}
     * @memberof BookInfoHanmoto
     */
    zaiko?: number;
    /**
     * 版
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    han?: string;
    /**
     * TRC用ジャンルコード
     * @type {number}
     * @memberof BookInfoHanmoto
     */
    genrecodetrc?: number;
    /**
     * TRC用児童ジャンルコード
     * @type {number}
     * @memberof BookInfoHanmoto
     */
    genrecodetrcjidou?: number;
    /**
     * TRC用備考
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    bikoutrc?: string;
    /**
     * NDC(日本十進分類法)
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    ndccode?: string;
    /**
     * 105文字程度の解説
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    kaisetsu105w?: string;
    /**
     * 版元からひとこと
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    hanmotokarahitokoto?: string;
    /**
     * 刊行形態
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    kankoukeitai?: string;
    /**
     * ラノベフラグ
     * @type {boolean}
     * @memberof BookInfoHanmoto
     */
    lanove?: boolean;
    /**
     * まえがきなど
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    maegakinado?: string;
    /**
     * 試し読みフラグ
     * @type {boolean}
     * @memberof BookInfoHanmoto
     */
    hastameshiyomi?: boolean;
    /**
     * 絶版日
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    datezeppan?: string;
    /**
     * 受賞情報
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    jushoujouhou?: string;
    /**
     * 重版予定日
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    datejuuhanyotei?: string;
    /**
     * 原書名
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    genshomei?: string;
    /**
     * 関連書
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    kanrensho?: string;
    /**
     * 追記
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    tsuiki?: string;
    /**
     * 帯の内容
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    obinaiyou?: string;
    /**
     * ルビの有無
     * @type {boolean}
     * @memberof BookInfoHanmoto
     */
    rubynoumu?: boolean;
    /**
     * 付録その他
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    furokusonota?: string;
    /**
     * 付録コード
     * @type {number}
     * @memberof BookInfoHanmoto
     */
    furoku?: number;
    /**
     * 雑誌コード
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    zasshicode?: string;
    /**
     * 類書・競合書
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    ruishokyougousho?: string;
    /**
     * その他の特記事項
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    sonotatokkijikou?: string;
    /**
     * 月号・号数
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    gatsugougousuu?: string;
    /**
     * 読者書き込み
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    dokushakakikomi?: string;
    /**
     * 関連書ISBN
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    kanrenshoisbn?: string;
    /**
     * 別称資料?
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    bessoushiryou?: string;
    /**
     * 重版情報
     * @type {Array<BookInfoHanmotoJyuhan>}
     * @memberof BookInfoHanmoto
     */
    jyuhan?: Array<BookInfoHanmotoJyuhan>;
    /**
     * 著者補足情報
     * @type {Array<BookInfoHanmotoAuthor>}
     * @memberof BookInfoHanmoto
     */
    author?: Array<BookInfoHanmotoAuthor>;
    /**
     * 書評情報
     * @type {Array<BookInfoHanmotoReviews>}
     * @memberof BookInfoHanmoto
     */
    reviews?: Array<BookInfoHanmotoReviews>;
    /**
     * 
     * @type {BookInfoHanmotoHanmotoinfo}
     * @memberof BookInfoHanmoto
     */
    hanmotoinfo?: BookInfoHanmotoHanmotoinfo;
    /**
     * 発売元
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    hatsubai?: string;
    /**
     * 発売元よみ
     * @type {string}
     * @memberof BookInfoHanmoto
     */
    hatsubaiyomi?: string;
}
/**
 * 
 * @export
 * @interface BookInfoHanmotoAuthor
 */
export interface BookInfoHanmotoAuthor {
    /**
     * 独自区分
     * @type {string}
     * @memberof BookInfoHanmotoAuthor
     */
    dokujikubun: string;
    /**
     * リストの番号
     * @type {number}
     * @memberof BookInfoHanmotoAuthor
     */
    listseq: number;
}
// ...
export interface BookInfoOnix {
    /**
     * 
     * @type {BookInfoOnixProductIdentifier}
     * @memberof BookInfoOnix
     */
    productIdentifier?: BookInfoOnixProductIdentifier;
    /**
     * ISBN（13桁）
     * @type {string}
     * @memberof BookInfoOnix
     */
    recordReference?: string;
    /**
     * 
     * @type {BookInfoOnixPublishingDetail}
     * @memberof BookInfoOnix
     */
    publishingDetail?: BookInfoOnixPublishingDetail;
    /**
     * 削除フラグ
     * @type {string}
     * @memberof BookInfoOnix
     */
    notificationType?: string;
    /**
     * 
     * @type {BookInfoOnixProductSupply}
     * @memberof BookInfoOnix
     */
    productSupply?: BookInfoOnixProductSupply;
    /**
     * 
     * @type {BookInfoOnixDescriptiveDetail}
     * @memberof BookInfoOnix
     */
    descriptiveDetail?: BookInfoOnixDescriptiveDetail;
    /**
     * 
     * @type {BookInfoOnixCollateralDetail}
     * @memberof BookInfoOnix
     */
    collateralDetail?: BookInfoOnixCollateralDetail;
}
// ...
export interface BookInfoOnixDescriptiveDetail {
    /**
     * 表記に利用する言語
     * @type {Array<BookInfoOnixDescriptiveDetailLanguage>}
     * @memberof BookInfoOnixDescriptiveDetail
     */
    language?: Array<BookInfoOnixDescriptiveDetailLanguage>;
    /**
     * 
     * @type {BookInfoOnixDescriptiveDetailTitleDetail}
     * @memberof BookInfoOnixDescriptiveDetail
     */
    titleDetail?: BookInfoOnixDescriptiveDetailTitleDetail;
    /**
     * 判型
     * @type {string}
     * @memberof BookInfoOnixDescriptiveDetail
     */
    productForm?: string;
    /**
     * 判型の詳細コード
     * @type {string}
     * @memberof BookInfoOnixDescriptiveDetail
     */
    productFormDetail?: string;
    /**
     * 
     * @type {BookInfoOnixDescriptiveDetailCollection}
     * @memberof BookInfoOnixDescriptiveDetail
     */
    collection?: BookInfoOnixDescriptiveDetailCollection;
    /**
     * 商品に関連する（数値的）範囲、程度など
     * @type {Array<BookInfoOnixDescriptiveDetailExtent>}
     * @memberof BookInfoOnixDescriptiveDetail
     */
    extent?: Array<BookInfoOnixDescriptiveDetailExtent>;
    /**
     * タテ・ヨコの実寸
     * @type {Array<BookInfoOnixDescriptiveDetailMeasure>}
     * @memberof BookInfoOnixDescriptiveDetail
     */
    measure?: Array<BookInfoOnixDescriptiveDetailMeasure>;
    /**
     * 商品の著者情報
     * @type {Array<BookInfoOnixDescriptiveDetailContributor>}
     * @memberof BookInfoOnixDescriptiveDetail
     */
    contributor?: Array<BookInfoOnixDescriptiveDetailContributor>;
    /**
     * セット商品分売可否
     * @type {string}
     * @memberof BookInfoOnixDescriptiveDetail
     */
    productComposition?: string;
    /**
     * 商品の対象読者を示す方法
     * @type {Array<BookInfoOnixDescriptiveDetailAudience>}
     * @memberof BookInfoOnixDescriptiveDetail
     */
    audience?: Array<BookInfoOnixDescriptiveDetailAudience>;
    /**
     * 商品の主題、テーマ、カテゴリなど
     * @type {Array<BookInfoOnixDescriptiveDetailSubject>}
     * @memberof BookInfoOnixDescriptiveDetail
     */
    subject?: Array<BookInfoOnixDescriptiveDetailSubject>;
    /**
     * 付録の有無
     * @type {Array<BookInfoOnixDescriptiveDetailProductPart>}
     * @memberof BookInfoOnixDescriptiveDetail
     */
    productPart?: Array<BookInfoOnixDescriptiveDetailProductPart>;
}
// ...
export interface BookInfoOnixProductIdentifier {
    /**
     * IDの種類[ISBN固定]
     * @type {string}
     * @memberof BookInfoOnixProductIdentifier
     */
    productIDType?: string;
    /**
     * ISBN（13桁）
     * @type {string}
     * @memberof BookInfoOnixProductIdentifier
     */
    iDValue?: string;
}
/**
 * 市場における商品の出版状況、供給等
 * @export
 * @interface BookInfoOnixProductSupply
 */
export interface BookInfoOnixProductSupply {
    /**
     * 
     * @type {BookInfoOnixProductSupplySupplyDetail}
     * @memberof BookInfoOnixProductSupply
     */
    supplyDetail?: BookInfoOnixProductSupplySupplyDetail;
}
// ...
export interface BookInfoOnixPublishingDetail {
    /**
     * 商品の出版に関する日付情報
     * @type {Array<BookInfoOnixPublishingDetailPublishingDate>}
     * @memberof BookInfoOnixPublishingDetail
     */
    publishingDate?: Array<BookInfoOnixPublishingDetailPublishingDate>;
    /**
     * 
     * @type {BookInfoOnixPublishingDetailImprint}
     * @memberof BookInfoOnixPublishingDetail
     */
    imprint?: BookInfoOnixPublishingDetailImprint;
    /**
     * 
     * @type {BookInfoOnixPublishingDetailPublisher}
     * @memberof BookInfoOnixPublishingDetail
     */
    publisher?: BookInfoOnixPublishingDetailPublisher;
}
// ...
export interface BookInfoSummary {
    /**
     * ISBN
     * @type {string}
     * @memberof BookInfoSummary
     */
    isbn: string;
    /**
     * 書名
     * @type {string}
     * @memberof BookInfoSummary
     */
    title: string;
    /**
     * 巻号
     * @type {string}
     * @memberof BookInfoSummary
     */
    volume: string;
    /**
     * シリーズ名
     * @type {string}
     * @memberof BookInfoSummary
     */
    series: string;
    /**
     * 著者名
     * @type {string}
     * @memberof BookInfoSummary
     */
    author: string;
    /**
     * 出版者
     * @type {string}
     * @memberof BookInfoSummary
     */
    publisher: string;
    /**
     * 出版年月
     * @type {string}
     * @memberof BookInfoSummary
     */
    pubdate: string;
    /**
     * 書影URL
     * @type {string}
     * @memberof BookInfoSummary
     */
    cover: string;
}
// ...
export class DefaultApi extends BaseAPI {
    /**
     * 書誌の概要を取得する。
     * @summary Returns book info summry
     * @param {Array<string>} isbn ISBNを指定。複数指定の場合はカンマ[,]で区切る。ISBNをXXX-X-XXXX-XXXX-Xの形式で指定する。1,000件まで複数指定。
     * @param {boolean} [pretty] 指定された場合、JSONを整形表示する。
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public getGet(isbn: Array<string>, pretty?: boolean, options?: any) {
        return DefaultApiFp(this.configuration).getGet(isbn, pretty, options)(this.axios, this.basePath);
    }

    /**
     * 書誌の概要を取得する。
     * @summary Returns book info summry
     * @param {Array<string>} isbn ISBNを指定。複数指定の場合はカンマ[,]で区切る。ISBNをXXX-X-XXXX-XXXX-Xの形式で指定する。1,0000件まで複数指定。
     * @param {boolean} [pretty] 指定された場合、JSONを整形表示する。
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public getPost(isbn: Array<string>, pretty?: boolean, options?: any) {
        return DefaultApiFp(this.configuration).getPost(isbn, pretty, options)(this.axios, this.basePath);
    }

}